[front/agent_builder/observability] - refactor: centralize user message origin data and update utils
 - Migrated USER_MESSAGE_ORIGIN_LABELS to observability constants to centralize label and color definitions
 - Updated `getSourceColor` and `buildSourceChartData` to reference the user message origin directly
 - Removed duplicate USER_MESSAGE_ORIGIN_LABELS and isUserMessageOrigin from assistant conversation types
 - Adjusted imports in notification activity to use the centralized user message origin checker function
