+++ front/types/assistant/conversation.ts
@@ -74,74 +74,6 @@ export type UserMessageOrigin =
   | "agent_handover"
   | "transcript";
 
-export const USER_MESSAGE_ORIGIN_LABELS: Record<
-  UserMessageOrigin,
-  { label: string; color: string }
-> = {
-  "github-copilot-chat": {
-    label: "GitHub Copilot Chat",
-    color: "bg-blue-500 dark:bg-blue-500-night",
-  },
-  agent_handover: {
-    label: "Agent handover",
-    color: "bg-gray-300 dark:bg-gray-300-night",
-  },
-  api: { label: "API", color: "bg-blue-300 dark:bg-blue-300-night" },
-  email: { label: "Email", color: "bg-violet-300 dark:bg-violet-300-night" },
-  excel: { label: "Excel", color: "bg-rose-300 dark:bg-rose-300-night" },
-  extension: {
-    label: "Chrome extension",
-    color: "bg-golden-300 dark:bg-golden-300-night",
-  },
-  gsheet: {
-    label: "Google Sheets",
-    color: "bg-green-300 dark:bg-green-300-night",
-  },
-  make: { label: "Make", color: "bg-gray-600 dark:bg-gray-600-night" },
-  n8n: { label: "n8n", color: "bg-success-muted dark:bg-success-muted-night" },
-  powerpoint: {
-    label: "PowerPoint",
-    color: "bg-violet-500 dark:bg-violet-500-night",
-  },
-  raycast: {
-    label: "Raycast",
-    color: "bg-rose-500 dark:bg-rose-500-night",
-  },
-  run_agent: {
-    label: "Sub-agent",
-    color: "bg-golden-500 dark:bg-golden-500-night",
-  },
-  slack: { label: "Slack", color: "bg-green-500 dark:bg-green-500-night" },
-  teams: {
-    label: "Teams",
-    color: "bg-highlight-muted dark:bg-highlight-muted-night",
-  },
-  transcript: {
-    label: "Transcript",
-    color: "bg-warning-muted dark:bg-warning-muted-night",
-  },
-  triggered_programmatic: {
-    label: "Trigger",
-    color: "bg-info-muted dark:bg-info-muted-night",
-  },
-  triggered: {
-    label: "Trigger",
-    color: "bg-info-muted dark:bg-info-muted-night",
-  },
-  web: { label: "Conversation", color: "bg-blue-500 dark:bg-blue-500-night" },
-  zapier: { label: "Zapier", color: "bg-blue-800 dark:bg-blue-800-night" },
-  zendesk: {
-    label: "Zendesk",
-    color: "bg-blue-800 dark:bg-blue-800-night",
-  },
-};
-
-export function isUserMessageOrigin(
-  origin?: string | null
-): origin is UserMessageOrigin {
-  return !!origin && origin in USER_MESSAGE_ORIGIN_LABELS;
-}
-
 export type UserMessageContext = {
   username: string;
   timezone: string;
